Address SonarCloud issues (#3666)
- Remove the assertions which can't fail in dropwizard-example.
- Split `ZipExceptionHandlingInputStreamTest` into two test classes. Since one of the inner classes was a parameterised test, it needed a static field and so the alternative was to reconfigure surefire. That seemed like a bit of a heavy-handed option so I just split the test up instead. I changed the `@Test` annotations to `@ParameterizedTest` as suggested by IntelliJ.
